Change interpolator to fast bilinear
Of the interpolators used by FFMpeg, only AREA, FAST_BILINEAR,
and POINT worked to resize the sample input to 8x8. Among these,
FAST_BILINEAR and POINT generated the most collisions (which
is good in this case). FAST_BILINEAR was chosen as a tradeoff
between information loss and modification recognition. Further
results might point in a different direction, but this works
for now (with scaling as the only transformation).